+++ react/components/EditorContainer/Sidebar/BlockEditor/BlockConfigurationEditor/index.tsx
@@ -3,7 +3,15 @@ import { JSONSchema6 } from 'json-schema'
 import React, { useMemo } from 'react'
 import { defineMessages, FormattedMessage, useIntl } from 'react-intl'
 import { FormProps } from 'react-jsonschema-form'
-import { Button, ToastConsumerFunctions, Toggle } from 'vtex.styleguide'
+// The styleguide is not exporting the warning icon
+import {
+  Button,
+  ToastConsumerFunctions,
+  Toggle,
+  // eslint-disable-next-line @typescript-eslint/ban-ts-ignore
+  // @ts-ignore
+  IconWarning,
+} from 'vtex.styleguide'
 
 import { useEditorContext } from '../../../../EditorContext'
 import EditableText from '../../../EditableText'
@@ -17,6 +25,12 @@ import Form from './Form'
 import { useComponentFormStateStack } from './hooks'
 import styles from './styles.css'
 import { getSchemas } from './utils'
+
+export enum ConfigurationStatus {
+  ACTIVE = 'ACTIVE',
+  INACTIVE = 'INACTIVE',
+  SCHEDULED = 'SCHEDULED',
+}
 
 interface Props {
   status?: string
@@ -40,7 +54,7 @@ interface Props {
   onSave: () => void
   showToast: ToastConsumerFunctions['showToast']
   title: ComponentSchema['title']
-  contentStatusFromRuntime: boolean
+  extensionStatus: ConfigurationStatus
 }
 
 const messages = defineMessages({
@@ -76,11 +90,11 @@ const BlockConfigurationEditor: React.FunctionComponent<Props> = ({
   onSave,
   showToast,
   title,
-  contentStatusFromRuntime,
+  extensionStatus,
 }) => {
   const intl = useIntl()
-  const statusToggleChecked =
-    status !== undefined ? status === 'active' : contentStatusFromRuntime
+  const toggleStatus = status ?? extensionStatus
+  const toggleChecked = toggleStatus === ConfigurationStatus.ACTIVE
 
   React.useEffect(() => {
     const { pageContext } = iframeRuntime.route
@@ -162,6 +176,18 @@ const BlockConfigurationEditor: React.FunctionComponent<Props> = ({
     return defaultContent && defaultContent.contentId === editingContentId
   }, [editor.blockData, editingContentId])
 
+  const getStatusWarning = () => {
+    if (extensionStatus === ConfigurationStatus.ACTIVE) {
+      return 'You can only activate content. Choose the one you want to be active and all others will be deactivated.'
+    } else if (condition?.statements.length) {
+      return 'This content will activate automatically on the start date you selected. '
+    } else if (toggleChecked) {
+      return 'If you create this content as active, you will see it in the preview now. Other content will be deactivated.'
+    }
+
+    return "If you create this content as inactive, you won't see it in the preview. You will see the currently active content."
+  }
+
   return (
     <div className="w-100 h-100 absolute flex flex-column">
       <div
@@ -223,7 +249,7 @@ const BlockConfigurationEditor: React.FunctionComponent<Props> = ({
               isSitewide={isSitewide}
               onConditionChange={onConditionChange}
               pageContext={iframeRuntime.route.pageContext}
-              contentStatusFromRuntime={contentStatusFromRuntime}
+              extensionStatus={extensionStatus}
             />
           )}
 
@@ -234,14 +260,26 @@ const BlockConfigurationEditor: React.FunctionComponent<Props> = ({
                 <FormattedMessage id="admin/pages.editor.components.status.title"></FormattedMessage>
               </div>
 
-              <div className="mv7">
+              <div className="mv5">
+                <div className="dib mb6">
+                  <IconWarning />
+                  <span className="ml3">{getStatusWarning()}</span>
+                </div>
+
                 <Toggle
-                  label={statusToggleChecked ? 'Activated' : 'Desactivated'}
+                  label={
+                    toggleChecked ? (
+                      <FormattedMessage id="admin/pages.editor.components.status.activationToggle.checked"></FormattedMessage>
+                    ) : (
+                      <FormattedMessage id="admin/pages.editor.components.status.activationToggle.unchecked"></FormattedMessage>
+                    )
+                  }
                   semantic
                   disabled={
-                    contentStatusFromRuntime || condition?.statements.length
+                    extensionStatus === ConfigurationStatus.ACTIVE ||
+                    condition?.statements.length
                   }
-                  checked={statusToggleChecked}
+                  checked={toggleChecked}
                   onChange={() => onStatusChange()}
                 />
               </div>